Adding day 23
I did not strictly follow the assignment.  I wanted a 2 player game with cars going in either direction.  The hit box for the cars could also be improved (turtles can be partially run over without dying), but I left it alone to make the game easier for my children to play.  Also, the game ends when a score of  5 is reached versus when a player is hit.  So this repo isn't the best example of what to do if you are following the course.